src/main/java/baekjoon/bronze/No10817.java
package baekjoon.bronze;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStreamReader;
import java.util.Arrays;
import java.util.StringTokenizer;

public class No10817 {
    public static void main(String[] args) throws IOException {
        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));

        int solution = solution(br.readLine());
        System.out.println(solution);
    }

    public static int solution(String input) {
        StringTokenizer st = new StringTokenizer(input);

        int[] arr = new int[3];
        for (int i = 0; i < 3; i++) {
            arr[i] = Integer.parseInt(st.nextToken());
        }
        Arrays.sort(arr);

        return arr[1];
    }
}
